Crawler updated
Added: waitpid (zombie bug fix)
Added: Send new found uri from http head to master
Updated: better variable names

+++ source/crawler/main.cpp
@@ -59,7 +59,7 @@ int main(int argc, char** argv)
 			}
 		}
 
-		Crawler crawler(ipaddress, port);
+		Crawler crawler(ipaddress, (unsigned short) port);
 
 		while (true)
 		{
